+++ data/model/src/main/java/com/englizya/model/request/Ticket.kt
@@ -17,6 +17,6 @@ data class Ticket(
     var ticketCategory: Int,
     var manifestoId: Int,
     var manifestoYear: Int,
-    var ticketLongitude: Int? = null,
-    var ticketLatitude: Int? = null,
+    var ticketLongitude: Double? = null,
+    var ticketLatitude: Double? = null,
 )
